requirements.txt
alembic==1.9.0
Brotli==1.0.9
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Compress==1.13
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-Minify==0.41
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
htmlmin==0.1.12
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
lesscpy==0.15.1
Mako==1.2.4
MarkupSafe==2.1.1
ply==3.11
python-dotenv==0.21.0
pytz==2022.6
rcssmin==1.1.1
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
xxhash==3.1.0
yapf==0.32.0
zipp==3.11.0
